Add demo for an indicator implementation with Fyne
This demo uses Fyne, which is supposed to be a cross-platform GUI
library, to create a pop-up indicator that shows up on top and
disappears by itself after a while.

As noted in the source code, there are limitations with Fyne that
potentially cannot be bypassed. We might have to look for something else
to implement the indicator-icon combo.